# [V1][Core] Support offloading KV cache to CPU.