## TL;DR
In V1, swap GPU KV cache blocks to CPU as they are evicted, and swap them back in if there is a cache hit.

## Swap Timing
CPU->GPU swap-in happen as requests get cache hits. This is a natural choice unless we want to do some kind of prediction.
GPU->CPU swap-out can be:
1. eagerly: as soon as a request is done and its blocks are freed
2. lazily: as we evict the GPU block when try to schedule requests
This PR chooses (2) hoping to minimize necessarily swaps but the downside is the swap-out overhead might be exposed.

In theory, an optimal solution would be to opportunistically swap out X cache blocks in certain cadence asynchronously (e.g. hidden behind the main CUDA graph) and maintain a free GPU block headroom. This will add more complexity and is leaved for future work.

## Implementation
This PR builds on top of the excellent V1 KV cache manager and blend in naturally with the current interface. The newly added metadata states: `cpu_block_pool` and `cached_block_hash_to_cpu_block` mirrors the GPU counter parts.
At a high-level, the KV cache manager accumulates swap-in/out decisions during a scheduling cycle, "flush" the swap decisions to the scheduler output so that model runners can issue the aggregated swap block calls before model execution. This minimizes the dispatch overhead.

For code simplicity, we avoid threading the scheduler output into various calls of kv cache manager, instead we accumulate those information in a few step_* fields (e.g. step_h2d_swap_map) and added a `end_schedule_step` callback to reset them after each schedule iteration. (Open to other suggestions.)

We reuse the user-facing flag "--swap-space" (not used in V1 atm) to determine the number cpu blocks. Though whether we want to change the default (4GB) is up for discussion.

## Benchmark
Initial benchmark result. Its performance ofc depends on the benchmark setup. Here I show two set of results, a top-line setup where workload should have many CPU cache hit and we want to see speedup, and a bottom-line where workload will have zero cache hit (but we still do GPU->CPU swaps) and we want to see how much slowdown it adds. Current conclusion is the swap out overhead is small given the bottomline benchmark result.

Topline
```
# Baseline with no CPU offloading: **14.2673 secs**
VLLM_USE_V1=1 python benchmark_long_document_qa_throughput.py --model meta-llama/Llama-3.1-8B --enable-prefix-caching --num-documents 10 --repeat-count 2 --repeat-mode tile --swap-space 0 --num-gpu-blocks-override 1252 --max-model-len 20010 --max-num-batched-tokens 20010

# With CPU offloading: **8.0952 secs**
VLLM_USE_V1=1 python benchmark_long_document_qa_throughput.py --model meta-llama/Llama-3.1-8B --enable-prefix-caching --num-documents 10 --repeat-count 2 --repeat-mode tile --swap-space 40 --num-gpu-blocks-override 1252 --max-model-len 20010 --max-num-batched-tokens 20010
```

Bottomline
```
# Baseline with no CPU offloading: **14.2608 secs**
VLLM_USE_V1=1 python benchmark_long_document_qa_throughput.py --model meta-llama/Llama-3.1-8B --enable-prefix-caching --num-documents 10 --repeat-count 2 --repeat-mode tile --swap-space 0 --num-gpu-blocks-override 1252 --max-model-len 20010 --max-num-batched-tokens 20010

# With CPU offloading: **14.2857 secs**
VLLM_USE_V1=1 python benchmark_long_document_qa_throughput.py --model meta-llama/Llama-3.1-8B --enable-prefix-caching --num-documents 10 --repeat-count 2 --repeat-mode tile --swap-space 4 --num-gpu-blocks-override 1252 --max-model-len 20010 --max-num-batched-tokens 20010
```

## TODO
- [ ] write tests
- [ ] profiling
- [ ] docs