@mengzhu28 2025-04-07
@WoosukKwon as discussed offline, created RFC #16144